Readme review fixes

+++ Drivers/SYSTEM/delay/delay.h
@@ -4,7 +4,7 @@
 #include "SYSTEM/sys/sys.h"
 
 void delay_init(uint16_t sysclk);
-void delay_ms(uint16_t nms);
+void delay_ms(uint32_t nms);
 void delay_us(uint32_t nus);
 
 void HAL_Delay(uint32_t Delay);
